move docLinks and yieldTokens to constants
