+++ features/manage-position/Create.tsx
@@ -24,7 +24,7 @@ import Connection from "../../containers/Connection";
 import { legacyEMPs } from "../../utils/legacyEmps";
 import { getLiquidationPrice } from "../../utils/getLiquidationPrice";
 import { isPricefeedInvertedFromTokenSymbol } from "../../utils/getOffchainPrice";
-import { DOCS_MAP } from "../../utils/getDocLinks";
+import { DOCS_MAP } from "../../constants/docLinks";
 import { toWeiSafe } from "../../utils/convertToWeiSafely";
 
 const Important = styled(Typography)`
