+++ features/manage-position/Withdraw.tsx
@@ -22,7 +22,7 @@ import Etherscan from "../../containers/Etherscan";
 
 import { getLiquidationPrice } from "../../utils/getLiquidationPrice";
 import { isPricefeedInvertedFromTokenSymbol } from "../../utils/getOffchainPrice";
-import { DOCS_MAP } from "../../utils/getDocLinks";
+import { DOCS_MAP } from "../../constants/docLinks";
 import { toWeiSafe } from "../../utils/convertToWeiSafely";
 
 const Important = styled(Typography)`
